Convert all html files to use header and main templates.  Fix Login page to work for Facebook.

+++ finalProject.py
@@ -81,8 +81,10 @@ def gconnect():
         response.headers['Content-Type'] = 'application/json'
 
     # Store the access token in the session for later use.
+    login_session['provider'] = 'google'
     login_session['credentials'] = credentials
     login_session['gplus_id'] = gplus_id
+    #response = make_response(json.dumps('Successfully connected user.', 200))
 
     # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
@@ -94,6 +96,7 @@ def gconnect():
     login_session['picture'] = data["picture"]
     login_session['email'] = data["email"]
 
+    # Create a new user if they do not already exist
     user_id = getUserID(login_session['email'])
     if not user_id:
         user_id = createUser(login_session)
@@ -124,23 +127,107 @@ def gdisconnect():
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
 
-    if result['status'] == '200':
-        # Reset the user's session.
-        del login_session['credentials']
-        del login_session['gplus_id']
-        del login_session['username']
-        del login_session['email']
-        del login_session['picture']
-
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    else:
+    if result['status'] != '200':
         # For whatever reason, the given token was invalid.
         response = make_response(json.dumps('Failed to revoke token for '
             'given user.'), 400)
         response.headers['Content-Type'] = 'application/json'
         return response
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+
+    # Exchange client token for long-lived server-side token with GET
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_id']
+    app_secret = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/' \
+          'access_token?grant_type=fb_exchange_token&client_id=%s' \
+          '&client_secret=%s&fb_exchange_token=%s' % (app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.4/me"
+    # Strip expire tag from access token
+    token = result.split("&")[0]
+
+    url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    print "url sent for API access: %s" % url
+    print "API JSON results: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    stored_token = token.split("=")[1]
+    login_session['access_token'] = stored_token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.4/me/picture?%s' \
+          '&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # See if the user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += '" style="width:300px; height:300px; border-radius:150px; ' \
+              '-webkit-border-radius: 150px; -moz-border-radius: 150px;">'
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "You have been logged out..."
+
+
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['credentials']
+            del login_session['gplus_id']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have been successfully logged out.")
+        return redirect(url_for('showRestaurants'))
+    else:
+        flash("You were not logged in to begin with!")
+        redirect(url_for('showRestaurants'))
+
 
 #Making an API Endpoint (GET Request)
 @app.route('/restaurants/JSON')
@@ -164,8 +251,12 @@ def itemMenuJSON(restaurant_id, menu_id):
 @app.route('/restaurants')
 @app.route('/restaurants/')
 def showRestaurants():
-    restaurants = session.query(Restaurant).all()
-    return render_template('restaurants.html', restaurant = restaurants)
+    restaurants = session.query(Restaurant).order_by(Restaurant.name)
+    if 'username' not in login_session:
+        return render_template('publicrestaurants.html', restaurant = restaurants)
+    else:
+        #user = getUserInfo(login_session['user_id'])
+        return render_template('restaurants.html', restaurant = restaurants)
 
 # CREATE A NEW RESTAURANT
 @app.route('/restaurants/new/', methods=['GET','POST'])
@@ -206,6 +297,11 @@ def deleteRestaurant(restaurant_id):
     if 'username' not in login_session:
         return redirect('/login')
     deletedRest = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if deletedRest.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized" \
+               "to delete this restaurant.  Please create your own restaurant" \
+               "or contact the current restaurant owner.');}</script><body " \
+               "onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(deletedRest)
         session.commit()
@@ -221,7 +317,13 @@ def deleteRestaurant(restaurant_id):
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant.id)
-    return render_template('menu.html', restaurant = restaurant, items = items)
+    creator = getUserInfo(restaurant.user_id)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items = items,
+                               restaurant = restaurant, creator = creator)
+    else:
+        return render_template('menu.html', restaurant = restaurant, items = items,
+                           creator = creator)
 
 # CREATE NEW MENU ITEM
 @app.route('/restaurants/<int:restaurant_id>/menu/new/', methods=['GET','POST'])
